requirements.txt
pandas==2.1.4
openpyxl==3.1.2
beautifulsoup4==4.12.2
requests==2.31.0
PyYAML==6.0.1
lxml==4.9.3
